+++ src/effitor/handler/undo/index.ts
@@ -1,1 +1,2 @@
-export { useUndoEffector, commandUndoHandler } from './effector';
+export { useUndoEffector, getUndoStack } from './effector';
+export { cmdHandler } from './handler'